refactor: extract domain handlers from rpc.rs into focused modules
Decompose the monolithic rpc.rs (~3300 lines) into domain-specific
modules following the existing rpc_auth.rs/rpc_kanban.rs pattern:

- rpc_channel.rs: channel.post/read handlers + shell unescaping helpers
- rpc_task.rs: task CRUD handlers + model/channel mapping helpers
- rpc_session.rs: session attach/detach/list + target resolution
- rpc_status.rs: status handler + task/channel activity helpers

rpc.rs remains as a thin dispatch layer with connection handling,
ParamExt trait, require_str! macro, and coworker lifecycle handlers.

All 51 RPC unit tests pass across the new module structure.
